+++ ufabc_planner-backend/src/modules/accounts/dtos/AuthenticateUserDTO.ts
@@ -1,5 +1,5 @@
 import { IsEmail, IsNotEmpty, IsString } from 'class-validator';
-import { messages } from 'utils/errors/messages';
+import { messages } from 'infra/http/errors/messages';
 import { IAuthenticateUser } from './interfaces/IAuthenticateUser';
 
 export class AuthenticateUserDTO implements IAuthenticateUser {
