+++ ufabc_planner-backend/src/modules/accounts/dtos/interfaces/ICreateUser.ts
