+++ ufabc_planner-backend/src/modules/accounts/services/AuthenticateUserService.ts
@@ -1,24 +1,27 @@
 import { compare } from 'bcryptjs';
 import auth from 'config/auth';
+import { prisma } from 'infra/prisma/client';
 import { sign } from 'jsonwebtoken';
-import { injectable } from 'tsyringe';
-import { AppError } from 'utils/errors/AppError';
-import { prisma } from 'utils/prisma';
+import { inject, injectable } from 'tsyringe';
+import { AppError } from 'infra/http/errors/AppError';
 import { AuthenticateUserDTO } from '../dtos/AuthenticateUserDTO';
+import { IUsersRepository } from '../repositories/IUsersRepository';
 
 interface IResponse {
   token: string;
 }
+
 @injectable()
 export class AuthenticateUserService {
+  constructor(
+    @inject('PrismaUserRepository')
+    private usersRepository: IUsersRepository
+  ) {}
+
   async execute({ email, password }: AuthenticateUserDTO): Promise<IResponse> {
     const { expires_in_token, secret_token } = auth;
 
-    const user = await prisma.user.findUnique({
-      where: {
-        email: email,
-      },
-    });
+    const user = await this.usersRepository.findByEmail(email);
 
     if (!user) {
       throw new AppError('Senha ou e-mail incorreto');
